# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SenseiLink API specification — marketplace for parents to find Shadow Teachers, Special Educators, and Medical Specialists
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /verifications/identity:
    post:
      tags: [verifications]
      operationId: submitIdentityVerification
      summary: Submit identity document for verification
      description: Professionals upload Aadhar/passport/ID with DPDP consent.
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitIdentityVerificationBody"
      responses:
        "201":
          description: Verification submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityVerificationDocument"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags: [verifications]
      operationId: getMyIdentityVerification
      summary: Get current professional's identity verification status
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Identity verification record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityVerificationDocument"
        "404":
          description: No verification submitted yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /verifications/certifications:
    post:
      tags: [verifications]
      operationId: submitCertification
      summary: Upload a qualification/certification document
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitCertificationBody"
      responses:
        "201":
          description: Certification uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificationDocument"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags: [verifications]
      operationId: getMyCertifications
      summary: List current professional's certification documents
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of certifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CertificationDocument"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    IdentityVerificationDocument:
      type: object
      required: [id, professionalId, documentType, fileKey, status, submittedAt]
      properties:
        id:
          type: integer
        professionalId:
          type: integer
        documentType:
          type: string
          enum: [aadhar, passport, driving_licence, national_id]
        fileKey:
          type: string
        status:
          type: string
          enum: [pending, verified, rejected]
        dpdpConsent:
          type: boolean
        submittedAt:
          type: string
          format: date-time
    SubmitIdentityVerificationBody:
      type: object
      required: [documentType, fileKey, dpdpConsent]
      properties:
        documentType:
          type: string
          enum: [aadhar, passport, driving_licence, national_id]
        fileKey:
          type: string
          description: Object path returned by the upload endpoint.
        dpdpConsent:
          type: boolean
          description: User explicitly consents to DPDP/GDPR data processing.
    CertificationDocument:
      type: object
      required: [id, professionalId, documentType, fileKey, uploadedAt]
      properties:
        id:
          type: integer
        professionalId:
          type: integer
        documentType:
          type: string
        fileKey:
          type: string
        uploadedAt:
          type: string
          format: date-time
    SubmitCertificationBody:
      type: object
      required: [documentType, fileKey]
      properties:
        documentType:
          type: string
          description: e.g. degree, diploma, certificate, training
        fileKey:
          type: string
          description: Object path returned by the upload endpoint.
...